Support spear-cli preview function from other component.
This patch will make the 'spear-cli' to support preview feature.
But the `magic` implementation is based on local files, this magicInternal
is based in-memory files. So this patch separate these implementation.

This patch is implementation for the 'takeyari' projects('WORKBENCH'),
so we need to design for supporting these implementation on one module.Support